align RRD start times
* This changes the start time of newly created RRDs to "0" (1970-01-01 00:00:00 UTC).

The idea behind is that all RRDs should be "aligned.
Given how RRD works, especially how it interpolates data onto steps, the start time has an influence over which time periods the steps (and PDPs) actually span.

Till now, PNP creates the RRDs with a start time that was set to the time of the first sample of perfdata retrieved from Nagios/Icinga.
These occur at arbitrary times and cannot be aligned, but now even the steps of the RRDs are likely unaligned.

This makes it a bit more vague to comare the values from different sensors.

Giving all RRDs a common start time should solve this problem at, as far as I can see, no disadvantages (not even with respect to RRD performance).