Fix timeout in process_perfdata.pl to use what is in the config
It seems as though the timeout option that you can specify in process_perfdata.cfg never
actually gets used. The only timeout that can get used is what is specified as default in the
process_perfdata.pl script or what is put on the command line using -t.

This seems like a bug to me or maybe I am missing something. I'm seeing this
in the latest 0.6.19 release and am using the NPCD bulk mode setup.

This commit is an attempt to try and correct the issue by first taking the timeout
specified on the command line. If there is none specified on the command then use
what is in the configuration file. If a timeout was not specified in either of those
places then use what is configured as the default in the script.