Fix timeranges
In commit 25de3550 the start and end times used for timeranges was
subjected to pnp::clean(). This was to help clean/validate all input
parameters to prevent XSS attacks. The problem is pnp::clean() replaces
all slashes with underscores which later breaks the conversion when
running strtotime().

It should be safe to remove cleaning the start and end inputs since they
get sent to getTimeRange() which will throw an exception if the input
can't be converted using strtotime() is not is_numeric().