Change the default source NULL instead of zero
In commit 25de3550 the 'source' value was configured to get a default of
zero if no input for source was supplied. This changed the behaviour of
the popup controller for example and possibly others.

After this change, instead of showing all of the graphs for a 'Disk'
check for exmaple, only source zero is shown.

Previously, the 'source' value would be NULL if no input was supplied.
With the value as NULL the popup controller will show all sources unless
a source is supplied as a parameter.