Add ability to pass JSON when creating 'pages'
The 'pages' feature is awesome but the configuration files lack the
ability to by dynamic. There are times when you want to create 'pages'
on the fly. Passing a JSON string as the page parameter allows for this.

There are times when images can be broken for various reasons. Most
notably is when using the 'SINGLE' RRD storage type and a DS was added
or removed. When creating PDF reports a single broken image will cause
the entire report to fail. This might not be desirable for some so add a
'skip_error' option to the PDF controller to help avoid this.